release 3.1.6 source code for net

+++ Services/Moderation/V3/Model/AudioModerationResultDetail.cs
@@ -129,7 +129,7 @@ namespace HuaweiCloud.SDK.Moderation.V3.Model
         /// 音频片段开始时间
         /// </summary>
         [JsonProperty("start_time", NullValueHandling = NullValueHandling.Ignore)]
-        public int? StartTime { get; set; }
+        public float? StartTime { get; set; }
 
         /// <summary>
         /// 音频片段审核处理建议： block：包含敏感信息，不通过 review：需要人工复检
@@ -140,7 +140,7 @@ namespace HuaweiCloud.SDK.Moderation.V3.Model
         /// 音频片段结束时间
         /// </summary>
         [JsonProperty("end_time", NullValueHandling = NullValueHandling.Ignore)]
-        public int? EndTime { get; set; }
+        public float? EndTime { get; set; }
 
         /// <summary>
         /// 音频片段标签
